split cluster and advanced cluster common methods that will diverge
