+++ internal/service/advancedcluster/data_source_advanced_clusters.go
@@ -279,14 +279,14 @@ func flattenAdvancedClusters(ctx context.Context, conn *matlas.Client, connV2 *a
 		result := map[string]any{
 			"advanced_configuration":         flattenProcessArgs(processArgs),
 			"backup_enabled":                 clusters[i].BackupEnabled,
-			"bi_connector_config":            FlattenBiConnectorConfig(clusters[i].BiConnector),
+			"bi_connector_config":            flattenBiConnectorConfig(clusters[i].BiConnector),
 			"cluster_type":                   clusters[i].ClusterType,
 			"create_date":                    clusters[i].CreateDate,
-			"connection_strings":             FlattenConnectionStrings(clusters[i].ConnectionStrings),
+			"connection_strings":             flattenConnectionStrings(clusters[i].ConnectionStrings),
 			"disk_size_gb":                   clusters[i].DiskSizeGB,
 			"encryption_at_rest_provider":    clusters[i].EncryptionAtRestProvider,
-			"labels":                         FlattenLabels(clusters[i].Labels),
-			"tags":                           FlattenTags(&clusters[i].Tags),
+			"labels":                         flattenLabels(clusters[i].Labels),
+			"tags":                           flattenTags(&clusters[i].Tags),
 			"mongo_db_major_version":         clusters[i].MongoDBMajorVersion,
 			"mongo_db_version":               clusters[i].MongoDBVersion,
 			"name":                           clusters[i].Name,
