+++ internal/service/advancedcluster/model_advanced_cluster_test.go
