+++ internal/service/cluster/data_source_clusters.go
@@ -379,7 +379,7 @@ func flattenClusters(ctx context.Context, d *schema.ResourceData, conn *matlas.C
 			"backup_enabled":                          clusters[i].BackupEnabled,
 			"provider_backup_enabled":                 clusters[i].ProviderBackupEnabled,
 			"cluster_type":                            clusters[i].ClusterType,
-			"connection_strings":                      advancedcluster.FlattenConnectionStrings(clusters[i].ConnectionStrings),
+			"connection_strings":                      flattenConnectionStrings(clusters[i].ConnectionStrings),
 			"disk_size_gb":                            clusters[i].DiskSizeGB,
 			"encryption_at_rest_provider":             clusters[i].EncryptionAtRestProvider,
 			"mongo_db_major_version":                  clusters[i].MongoDBMajorVersion,
@@ -403,10 +403,10 @@ func flattenClusters(ctx context.Context, d *schema.ResourceData, conn *matlas.C
 			"provider_instance_size_name":                     clusters[i].ProviderSettings.InstanceSizeName,
 			"provider_name":                                   clusters[i].ProviderSettings.ProviderName,
 			"provider_region_name":                            clusters[i].ProviderSettings.RegionName,
-			"bi_connector_config":                             advancedcluster.FlattenBiConnectorConfig(clusters[i].BiConnector),
+			"bi_connector_config":                             flattenBiConnectorConfig(clusters[i].BiConnector),
 			"replication_specs":                               flattenReplicationSpecs(clusters[i].ReplicationSpecs),
-			"labels":                                          advancedcluster.FlattenLabels(clusters[i].Labels),
-			"tags":                                            advancedcluster.FlattenTags(clusters[i].Tags),
+			"labels":                                          flattenLabels(clusters[i].Labels),
+			"tags":                                            flattenTags(clusters[i].Tags),
 			"snapshot_backup_policy":                          snapshotBackupPolicy,
 			"termination_protection_enabled":                  clusters[i].TerminationProtectionEnabled,
 			"version_release_system":                          clusters[i].VersionReleaseSystem,
